+++ CHANGELOG.md
@@ -24,6 +24,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
  - #2534, Fix embedding a computed relationship with a normal relationship - @steve-chavez
  - #2362, Fix error message when [] is used inside select - @wolfgangwalther
  - #2475, Disallow !inner on computed columns - @wolfgangwalther
+ - #2285, Ignore leading and trailing spaces in column names when parsing the query string - @wolfgangwalther
 
 ### Changed